Refactored NearestNeighborRanker to work with N classes
